Split db_schema_classes into many files in new folder db_schema_classes in lib.
