+++ lib/create_project_form.dart
@@ -6,7 +6,8 @@ import 'package:p2bp_2025spring_mobile/project_map_creation.dart';
 import 'package:p2bp_2025spring_mobile/theme.dart';
 import 'package:p2bp_2025spring_mobile/widgets.dart';
 
-import 'db_schema_classes.dart';
+import 'db_schema_classes/member_class.dart';
+import 'db_schema_classes/team_class.dart';
 
 class CreateProjectForm extends StatefulWidget {
   final Member member;
