+++ lib/lighting_profile_test.dart
@@ -8,7 +8,8 @@ import 'package:p2bp_2025spring_mobile/extensions.dart';
 import 'package:p2bp_2025spring_mobile/theme.dart';
 import 'package:p2bp_2025spring_mobile/widgets.dart';
 
-import 'db_schema_classes.dart';
+import 'db_schema_classes/project_class.dart';
+import 'db_schema_classes/specific_test_classes/lighting_profile_test_class.dart';
 import 'google_maps_functions.dart';
 
 class LightingProfileTestPage extends StatefulWidget {
