+++ lib/manage_team_members_form.dart
@@ -4,7 +4,9 @@ import 'package:flutter/material.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:p2bp_2025spring_mobile/widgets.dart';
 
-import 'db_schema_classes.dart';
+import 'db_schema_classes/member_class.dart';
+import 'db_schema_classes/misc_class_stuff.dart';
+import 'db_schema_classes/team_class.dart';
 import 'theme.dart';
 
 class ManageTeamMembersForm extends StatefulWidget {
