+++ lib/project_map_creation.dart
@@ -9,7 +9,10 @@ import 'package:p2bp_2025spring_mobile/assets.dart';
 import 'package:p2bp_2025spring_mobile/theme.dart';
 import 'package:p2bp_2025spring_mobile/widgets.dart';
 
-import 'db_schema_classes.dart';
+import 'db_schema_classes/member_class.dart';
+import 'db_schema_classes/project_class.dart';
+import 'db_schema_classes/standing_point_class.dart';
+import 'db_schema_classes/team_class.dart';
 import 'google_maps_functions.dart';
 import 'home_screen.dart';
