+++ lib/team_settings_page.dart
@@ -10,7 +10,9 @@ import 'package:p2bp_2025spring_mobile/manage_team_members_form.dart';
 import 'package:p2bp_2025spring_mobile/widgets.dart';
 
 import 'create_project_form.dart';
-import 'db_schema_classes.dart';
+import 'db_schema_classes/member_class.dart';
+import 'db_schema_classes/project_class.dart';
+import 'db_schema_classes/team_class.dart';
 import 'project_details_page.dart';
 import 'theme.dart';
